Add a revised case design with internal nut slots
This is a revised version of the original case design that eliminates the nuts
on the back of the device in favour of nuts contained within a slot within the
case itself (inspired by a technique used on some Prusa printer parts).

Also removes the micro-USB slot on the side.