Ensure pio installs required dependencies.
The new post_build.py file from f8a731c requires esptool to be in the
environment, but pio doesn't appear to install this by default, and
vscode/ubuntu 23.04+ run everything in python venv's meaning that even
if you have it installed in your general workspace pio can't necessarily
find it.

This ensures that esptool is always available in the venv that pio is
running in.